Refactor domain과service 파일 구분
