+++ .idea/.name
@@ -1,1 +1,1 @@
-KIOSK_JDBC
+KioskApp.java
